refactor: refactor data types in Repositories

+++ src/Lab5/Lab5.Application.Abstractions/Repositories/IAdminAccountsRepository.cs
@@ -2,5 +2,5 @@
 
 public interface IAdminAccountsRepository
 {
-    string? FindAdminPasswordById(long userId, decimal accountId);
+    string? FindAdminPasswordById(long userId, long accountId);
 }
