+++ src/Lab5/Lab5.Application.Abstractions/Repositories/IOperationsHistoryRepository.cs
@@ -4,5 +4,5 @@ namespace Application.Abstractions.Repositories;
 
 public interface IOperationsHistoryRepository
 {
-    IEnumerable<Operation> FindOperationsHistoryById(decimal accountId);
+    IEnumerable<Operation> FindOperationsHistoryById(long accountId);
 }
