+++ src/Lab5/Lab5.Application.Models/Accounts/CustomerAccount.cs
@@ -1,7 +1,7 @@
 ﻿namespace Application.Models.Accounts;
 
 public record CustomerAccount(
-    decimal AccountId,
+    long AccountId,
     long UserId,
     decimal Balance,
     CustomerAccountState State,
