+++ src/Lab5/Lab5.Application.Models/Operations/Operation.cs
@@ -1,3 +1,3 @@
 ﻿namespace Application.Models.Operations;
 
-public record Operation(decimal AccountId, long OperationId, decimal Amount, OperationType Type, OperationState State, DateTime Date);
+public record Operation(long AccountId, long OperationId, decimal Amount, OperationType Type, OperationState State, DateTime Date);
